+++ src/mpservice/_http.py
@@ -7,11 +7,10 @@ import socket
 import sys
 import time
 import warnings
-from typing import Any, Callable, List, Optional
+from typing import Any, Callable, List, Optional, Union
 
 import click
 import uvicorn
-from asgiref.typing import ASGIApplication  # such as `starlette.applications.Starlette`
 
 from mpservice.multiprocessing import MP_SPAWN_CTX, SpawnProcess
 
@@ -193,7 +192,7 @@ UNSET = object()
 
 
 def start_server(
-    app: str | 'ASGIApplication',
+    app: Union[str, 'ASGIApplication'],
     *,
     host: str = "0.0.0.0",
     port: int = 8000,